Moved containers list to ActivityContent section
